added connections len to the connections object

+++ pkg/network/event_common.go
@@ -137,7 +137,7 @@ type Connections struct {
 	HTTP2                       map[http.Key]*http.RequestStats
 	Kafka                       map[kafka.Key]*kafka.RequestStat
 	DNSStats                    dns.StatsByKeyByNameByType
-	Length                      int32
+	Length                      int64
 }
 
 // NewConnections create a new Connections object
